requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.5.3

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.1.2

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Voice processing (for technician server)
sounddevice>=0.4.6
numpy>=1.26.3
scipy>=1.11.4

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0 